Structure sentry logging from TaskInstance
The previous implementation sent the same HTML formatted string as the
email notification. Sentry doesn't in any case render it as HTML, and
this doesn't provide metadata in a form sentry can usefully use.

The actual message is shortened to just (Retrying|Failed): DAG.TASK: EXCEPTION;
severity is now "Warning" for retries and "Error" for failed.
Most of what was previously folded into the HTML message is now included
as fields in the "extra" dict.